Do not automatically try to add 1 prefix to dice specs
When user input a dice-spec like d4 the parser tried to
be clever and turn that into 1d4. But that does not work
very well now with custom dice names and the automatic
attempt to parse strings as dice specifications
for dice-tables. For instance it thought that
dragon was a 1dRAGON dice (from the example-dragon table).