Adding custom dice for buckets-of-dice
New included custom dice B6, B5, B4, B3, B2,
that are standard d6 except for Bn only a roll of n or more
counts as a hit. Use for rolling buckets of dice
and counting number of hits.

Also added comments in default value for decide-custom-dice
variable to make it more readable