Avoid error for decide-dwim-insert on first line
When pressing ? ?, for decide-dwim-insert, on the first line of a buffer, the
backwards search for beginning of line failed, causing an error from
search-backward-regexp to be printed. Adding t for the optional NOERROR argument
to search-backward-regexp makes it silently return nil instead. Now
decide-dwim-insert works as expected on the first line.